model_serving/requirements.txt
fastapi>=0.95.0
uvicorn 
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]
pydantic[email]
SQLAlchemy>=2.0.22
pymysql>=1.1.0
scikit-learn>=1.3.1
pandas>=2.1.1
onnx>=1.14.1
onnxruntime-gpu>=1.16.0
PyYAML
python-dotenv

-e .